+++ src/main/java/com/onulstore/config/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.onulstore.jwt;
+package com.onulstore.config.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
